Removed duplicate classes from wit model
