--- src/main/java/de/refactoringbot/model/wit/JavaAnnotation.java
@@ -1,62 +1,0 @@
-package de.refactoringbot.model.wit;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class JavaAnnotation {
-
-	@JsonProperty("confidence")
-	private Double confidence;
-	@JsonProperty("value")
-	private String value;
-	@JsonProperty("type")
-	private String type;
-	@JsonIgnore
-	private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-	@JsonProperty("confidence")
-	public Double getConfidence() {
-		return confidence;
-	}
-
-	@JsonProperty("confidence")
-	public void setConfidence(Double confidence) {
-		this.confidence = confidence;
-	}
-
-	@JsonProperty("value")
-	public String getValue() {
-		return value;
-	}
-
-	@JsonProperty("value")
-	public void setValue(String value) {
-		this.value = value;
-	}
-
-	@JsonProperty("type")
-	public String getType() {
-		return type;
-	}
-
-	@JsonProperty("type")
-	public void setType(String type) {
-		this.type = type;
-	}
-
-	@JsonAnyGetter
-	public Map<String, Object> getAdditionalProperties() {
-		return this.additionalProperties;
-	}
-
-	@JsonAnySetter
-	public void setAdditionalProperty(String name, Object value) {
-		this.additionalProperties.put(name, value);
-	}
-
-}
